move timer start/stop handling to app level; create scramble card component
